Got the app to a stable location.
Turns out, I simply did not have the Dockerfile in the app build, I had it on the toplevel, when it should be with the process it is building.

For example, if you have a frontend and backend, each should have their own Dockerfile within their directory.

Docker compose goes at the top-level of the project.